moves NavBar to component global scope
